+++ tools/postprocessing_tool_json2dew_converter/lib_data_io_shapefile.py
@@ -3,8 +3,8 @@ Class Features
 
 Name:          lib_data_io_shapefile
 Author(s):     Fabio Delogu (fabio.delogu@cimafoundation.org)
-Date:          '20210225'
-Version:       '1.0.0'
+Date:          '20210730'
+Version:       '1.0.1'
 """
 
 #######################################################################################
@@ -23,52 +23,52 @@ logging.getLogger('geopandas').setLevel(logging.WARNING)
 
 
 # -------------------------------------------------------------------------------------
-# Method to find section data
-def find_data_section(section_df, section_name=None, basin_name=None,
-                      tag_column_section_in='section_name', tag_column_basin_in='section_domain',
-                      tag_column_section_out='section_name', tag_column_basin_out='basin_name'):
+# Method to find point data
+def find_data_point(point_df, point_name=None, basin_name=None,
+                    tag_column_point_in='point_name', tag_column_basin_in='point_domain',
+                    tag_column_point_out='point_name', tag_column_basin_out='basin_name'):
 
-    section_name_ref = section_name.lower()
+    point_name_ref = point_name.lower()
     basin_name_ref = basin_name.lower()
 
-    section_name_list = section_df[tag_column_section_in].values
-    basin_name_list = section_df[tag_column_basin_in].values
+    point_name_list = point_df[tag_column_point_in].values
+    basin_name_list = point_df[tag_column_basin_in].values
 
-    section_dict_tmp = {tag_column_section_in: section_name_list, tag_column_basin_in: basin_name_list}
-    section_df_tmp = pd.DataFrame(data=section_dict_tmp)
-    section_df_tmp = section_df_tmp.astype(str).apply(lambda x: x.str.lower())
+    point_dict_tmp = {tag_column_point_in: point_name_list, tag_column_basin_in: basin_name_list}
+    point_df_tmp = pd.DataFrame(data=point_dict_tmp)
+    point_df_tmp = point_df_tmp.astype(str).apply(lambda x: x.str.lower())
 
-    point_idx = section_df_tmp[(section_df_tmp[tag_column_section_in] == section_name_ref) &
-                                 (section_df_tmp[tag_column_basin_in] == basin_name_ref)].index
+    point_idx = point_df_tmp[(point_df_tmp[tag_column_point_in] == point_name_ref) &
+                             (point_df_tmp[tag_column_basin_in] == basin_name_ref)].index
 
     if point_idx.shape[0] == 1:
         point_idx = point_idx[0]
-        point_dict = section_df.iloc[point_idx, :].to_dict()
+        point_dict = point_df.iloc[point_idx, :].to_dict()
 
-        point_dict[tag_column_section_out] = point_dict.pop(tag_column_section_in)
+        point_dict[tag_column_point_out] = point_dict.pop(tag_column_point_in)
         point_dict[tag_column_basin_out] = point_dict.pop(tag_column_basin_in)
 
     elif point_idx.shape[0] == 0:
-        raise IOError('Section selection failed; section not found')
+        raise IOError('Point selection failed; point not found')
     else:
-        raise NotImplementedError('Section selection failed for unknown reason.')
+        raise NotImplementedError('Point selection failed for unknown reason.')
 
     return point_dict
 # -------------------------------------------------------------------------------------
 
 
 # -------------------------------------------------------------------------------------
-# Method to read section(s) data (shapefile)
-def read_data_section(file_name, columns_name_expected_in=None, columns_name_expected_out=None, columns_name_type=None):
+# Method to read dam(s) data (shapefile)
+def read_data_dam(file_name, columns_name_expected_in=None, columns_name_expected_out=None, columns_name_type=None):
 
     if columns_name_expected_in is None:
         columns_name_expected_in = [
-            'HMC_X', 'HMC_Y', 'LAT', 'LON', 'BASIN', 'SEC_NAME', 'SEC_RS', 'AREA', 'Q_THR1', 'Q_THR2']
+            'HMC_X', 'HMC_Y', 'LAT', 'LON', 'BASIN', 'NAME', 'STATION', 'AREA', 'Q_THR1', 'Q_THR2']
 
     if columns_name_expected_out is None:
         columns_name_expected_out = [
-            'hmc_idx_x', 'hmc_idx_y', 'latitude', 'longitude', 'section_domain', 'section_name', 'section_code',
-            'section_drained_area', 'section_discharge_thr_alert', 'section_discharge_thr_alarm']
+            'hmc_idx_x', 'hmc_idx_y', 'latitude', 'longitude', 'point_domain', 'point_name', 'point_code',
+            'point_drained_area', 'point_discharge_thr_alert', 'point_discharge_thr_alarm']
 
     if columns_name_type is None:
         columns_name_type = ['int', 'int', 'float', 'float',
@@ -77,7 +77,7 @@ def read_data_section(file_name, columns_name_expected_in=None, columns_name_exp
     file_dframe_raw = gpd.read_file(file_name)
     file_rows = file_dframe_raw.shape[0]
 
-    section_obj = {}
+    point_obj = {}
     for column_name_in, column_name_out, column_type in zip(columns_name_expected_in,
                                                             columns_name_expected_out, columns_name_type):
         if column_name_in in file_dframe_raw.columns:
@@ -92,9 +92,52 @@ def read_data_section(file_name, columns_name_expected_in=None, columns_name_exp
             else:
                 raise NotImplementedError('Datatype not implemented yet')
 
-        section_obj[column_name_out] = column_data
+        point_obj[column_name_out] = column_data
 
-    section_df = pd.DataFrame(data=section_obj)
+    point_df = pd.DataFrame(data=point_obj)
 
-    return section_df
+    return point_df
+# -------------------------------------------------------------------------------------
+
+
+# -------------------------------------------------------------------------------------
+# Method to read section(s) data (shapefile)
+def read_data_section(file_name, columns_name_expected_in=None, columns_name_expected_out=None, columns_name_type=None):
+
+    if columns_name_expected_in is None:
+        columns_name_expected_in = [
+            'HMC_X', 'HMC_Y', 'LAT', 'LON', 'BASIN', 'SEC_NAME', 'SEC_RS', 'AREA', 'Q_THR1', 'Q_THR2']
+
+    if columns_name_expected_out is None:
+        columns_name_expected_out = [
+            'hmc_idx_x', 'hmc_idx_y', 'latitude', 'longitude', 'point_domain', 'point_name', 'point_code',
+            'point_drained_area', 'point_discharge_thr_alert', 'point_discharge_thr_alarm']
+
+    if columns_name_type is None:
+        columns_name_type = ['int', 'int', 'float', 'float',
+                             'str', 'str', 'str', 'float', 'float', 'float']
+
+    file_dframe_raw = gpd.read_file(file_name)
+    file_rows = file_dframe_raw.shape[0]
+
+    point_obj = {}
+    for column_name_in, column_name_out, column_type in zip(columns_name_expected_in,
+                                                            columns_name_expected_out, columns_name_type):
+        if column_name_in in file_dframe_raw.columns:
+            column_data = file_dframe_raw[column_name_in].values.tolist()
+        else:
+            if column_type == 'int':
+                column_data = [-9999] * file_rows
+            elif column_type == 'str':
+                column_data = [''] * file_rows
+            elif column_type == 'float':
+                column_data = [-9999.0] * file_rows
+            else:
+                raise NotImplementedError('Datatype not implemented yet')
+
+        point_obj[column_name_out] = column_data
+
+    point_df = pd.DataFrame(data=point_obj)
+
+    return point_df
 # -------------------------------------------------------------------------------------